pyproject.toml
[tool.poetry]
name = "monefy-web-app"
version = "0.1.0"
description = ""
authors = ["Bohdan Solianyk <bohdan_solianyk@epam.com>"]
readme = "README.md"
packages = [{include = "monefy_web_app"}]

[tool.poetry.dependencies]
python = "^3.10"
sanic = "^22.9.0"
cryptography = "^38.0.1"
dropbox = "^11.35.0"
sanic-ext = "^22.9.0"
jinja2 = "^3.1.2"
pyjwt = "^2.5.0"
psycopg2-binary = "^2.9.5"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
pylint = "^2.15.4"
isort = "^5.10.1"
mypy = "^0.982"
pytest = "^7.1.3"
bandit = "^1.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.master]
output-format = "colorized"

[tool.pylint.messages_control]
disable = ["unused-argument", "logging-fstring-interpolation"]

[tool.pylint.design]
max-attributes = "20"
max-parents = "8"
max-args = "15"

[tool.pytest.ini_options]
asyncio_mode = "strict"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = "True"
check_untyped_defs = "True"
disallow_any_generics = "True"
disallow_incomplete_defs = "True"
disallow_untyped_defs = "True"
disallow_untyped_calls = "True"
no_implicit_optional = "True"
warn_redundant_casts = "True"
warn_unused_ignores = "True"
exclude = ["venv", "tests"]
